Voice chat & tabs rework, other improvements
The tabs had to be reworked as a workaround for Valve's broken drag behavior. Having multiple tabs open and with the window relatively small, the leftmost tab selected, attempting to drag the window instead picks up one of the tabs hidden behind the drag area. This is caused by the tabs being set to no-drag. It can result in dropping tabs onto conversations by mistake, which sends a blank message to the user.